requirements.txt
Flask==3.1.0
Flask_Admin==1.6.1
Flask_Bcrypt==1.0.1
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
Jinja2==3.1.5
piexif==1.1.3
Pillow==11.1.0
Werkzeug==3.1.3
